requirements.txt
attrs~=22.2.0
boto3
deepgram-sdk
python-telegram-bot[job_queue,rate-limiter]
python-dotenv~=0.21.0
aiogram~=2.24
pydantic
